refactored Trade page
