+++ src/page/Trade/Summary/index.tsx
@@ -1,22 +1,112 @@
 import { Box, Heading, Text, VStack } from "@chakra-ui/react"
 import SendTxButton from "./SendTxButton"
 import TxInfoTable from "./TxInfoTable"
+import { Trade } from "container/trade"
+import { Transaction } from "container/transaction"
+import { useCallback } from "react"
+import { usePositionSize } from "../usePositionSize"
+import { big2BigNum } from "util/format"
+import { Connection } from "container/connection"
+import { PerpdexExchangeContainer } from "container/perpdexExchangeContainer"
+import { PerpdexMarketContainer } from "container/perpdexMarketContainer"
+import Big from "big.js"
+
+function calcPositionSize(isBaseToQuote: boolean, notional: Big, markPrice: Big) {
+    const basePosition = isBaseToQuote ? notional.mul(markPrice) : notional
+    const oppositPosition = isBaseToQuote ? notional : notional.mul(markPrice)
+    return {
+        basePosition,
+        oppositPosition,
+    }
+}
 
 function Summary() {
+    const {
+        state: { currentMarketInfo, markPrice, contract },
+    } = PerpdexMarketContainer.useContainer()
+    const { slippage, isBaseToQuote, collateral, leverage } = Trade.useContainer()
+    const { openPosition } = PerpdexExchangeContainer.useContainer()
+    const { isLoading } = Transaction.useContainer()
+    const { positionSize, isCalculating } = usePositionSize()
+    const { account } = Connection.useContainer()
+
+    // const { size: openedSize, margin: openedMargin, unrealizedPnl, outputPrice } = useOpenedPositionSize("")
+
+    // /* prepare data for UI */
+    // const entryPrice: Big | null = useMemo(() => {
+    //     if (!isCalculating && positionSize !== "" && collateral !== null) {
+    //         const b_positionSize = new Big(positionSize)
+    //         if (b_positionSize.eq(0) || collateral.mul(leverage).eq(0)) {
+    //             return null
+    //         }
+    //         return collateral.mul(leverage).div(b_positionSize)
+    //     }
+    //     return null
+    // }, [collateral, isCalculating, leverage, positionSize])
+
+    const isDisabled = isLoading || isCalculating || collateral === null || collateral.eq(0)
+
+    const handleOnTrade = useCallback(async () => {
+        if (collateral && currentMarketInfo && markPrice && account) {
+            const isExactInput = isBaseToQuote
+            // const amount = new Big(positionSize)
+
+            const positions = calcPositionSize(isBaseToQuote, collateral, markPrice)
+
+            const _slippage = slippage / 100
+
+            let oppositeAmountBount
+            if (isExactInput) {
+                oppositeAmountBount = positions.oppositPosition.mul(1 - _slippage)
+            } else {
+                oppositeAmountBount = positions.oppositPosition.mul(1 + _slippage)
+            }
+
+            console.log(
+                "openPosition",
+                isBaseToQuote,
+                isExactInput,
+                positions.basePosition.toString(),
+                oppositeAmountBount.toString(),
+            )
+
+            // const results = await contractExecuter?.contract.callStatic.openPosition({
+            //     trader: account,
+            //     market: currentMarket.baseAddress,
+            //     isBaseToQuote,
+            //     isExactInput,
+            //     amount: big2BigNum(positions.basePosition),
+            //     oppositeAmountBound: big2BigNum(oppositeAmountBount),
+            //     deadline: BigNumber.from(2).pow(96)
+            // })
+
+            // if (results?.base && results?.quote) {
+            //     console.log(bigNum2Big(results?.base, 18).toString(), bigNum2Big(results?.quote, 18).toString())
+            // }
+
+            openPosition(
+                isBaseToQuote,
+                isExactInput,
+                big2BigNum(positions.basePosition),
+                big2BigNum(oppositeAmountBount),
+            )
+        }
+    }, [account, collateral, currentMarketInfo, isBaseToQuote, markPrice, openPosition, slippage])
+
     return (
         <>
             <Heading w="full" size="md">
                 Transaction Summary
             </Heading>
+            <Box width="100%" borderStyle="solid" borderWidth="1px" borderColor="gray.200" borderRadius="12px">
+                <TxInfoTable markPrice={markPrice} priceImpact="" estimateGasFee={Big(0)} fundingRatio="" />
+            </Box>
             <VStack spacing={2} width="full">
-                <SendTxButton />
+                <SendTxButton isDisabled={isDisabled} isLoading={isLoading} handleOnClick={handleOnTrade} />
                 <Text fontSize="sm" color="gray.500">
                     Confirm in Metamask
                 </Text>
             </VStack>
-            <Box width="100%" borderStyle="solid" borderWidth="1px" borderColor="gray.200" borderRadius="12px">
-                <TxInfoTable />
-            </Box>
         </>
     )
 }
