+++ src/page/Trade/TradeInput.tsx/index.tsx
@@ -3,12 +3,14 @@ import React from "react"
 import SideSwitcher from "./SideSwitcher"
 import Collateral from "./Collateral"
 import Slippage from "./Slippage"
+import Position from "./Position"
 
 function TradeInput() {
     return (
         <>
             <SideSwitcher />
             <Collateral />
+            <Position />
             <Slippage />
         </>
     )
